requirements.txt
python-telegram-bot==20.7
openai==1.3.0
python-dotenv==1.0.0
pysrt==1.1.2
aiofiles==23.2.0
asyncio==3.4.3